backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pyrogram==2.0.106
tgcrypto==1.2.5
motor==3.1.2
python-dotenv==1.0.0
cinemagoer==2023.5.1
pydantic==1.10.7
python-slugify==8.0.1
aiohttp==3.8.4